project_6/ex01/requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.7
click==8.1.7
coloredlogs==15.0.1
Deprecated==1.2.14
fastapi==0.115.0
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==5.0.0
monotonic==1.6
mpmath==1.3.0
narwhals==1.8.2
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
posthog==3.6.6
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.5
streamlit==1.38.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.0.1
wrapt==1.16.0
zipp==3.20.2
